Refactor: お気に入り機能を分離
